backend/requirements.txt
uWSGI==2.0.19.1
falcon==1.0.0
falcon-cors==1.1.2
PyYAML==5.3.1
bcrypt==3.1.0
pydantic==1.8.1

psycopg2==2.8.6
SQLAlchemy==1.4.10
sqlalchemy-stubs==0.4
alembic==0.8.7

hiredis==0.2.0
redis==2.10.5

guess-language-spirit==0.5.2
pyenchant==1.6.8

